requirements.txt
requests
python-dotenv
boto3
google-genai
openai
